+++ classes/BlueChip/Security/Modules/Checklist/Checks/DisplayOfPhpErrorsIsOff.php
@@ -9,17 +9,20 @@ use BlueChip\Security\Modules\Checklist;
 
 class DisplayOfPhpErrorsIsOff extends Checklist\BasicCheck
 {
-    public function __construct()
+    public function getDescription(): string
     {
-        parent::__construct(
-            __('Display of PHP errors is off', 'bc-security'),
-            \sprintf(
-                /* translators: 1: link to PHP manual documentation on display-errors php.ini setting, 2: link to WordPress Handbook article */
-                esc_html__('%1$s to the screen as part of the output on live system. In WordPress environment, %2$s when directly loading certain files.', 'bc-security'),
-                '<a href="' . esc_url(__('https://www.php.net/manual/en/errorfunc.configuration.php#ini.display-errors', 'bc-security')) . '" rel="noreferrer">' . esc_html__('Errors should never be printed', 'bc-security') . '</a>',
-                '<a href="' . esc_url(__('https://make.wordpress.org/core/handbook/testing/reporting-security-vulnerabilities/#why-are-there-path-disclosures-when-directly-loading-certain-files', 'bc-security')) . '" rel="noreferrer">' . esc_html__('display of errors can lead to path disclosures', 'bc-security') . '</a>'
-            )
+        return \sprintf(
+            /* translators: 1: link to PHP manual documentation on display-errors php.ini setting, 2: link to WordPress Handbook article */
+            esc_html__('%1$s to the screen as part of the output on live system. In WordPress environment, %2$s when directly loading certain files.', 'bc-security'),
+            '<a href="' . esc_url(__('https://www.php.net/manual/en/errorfunc.configuration.php#ini.display-errors', 'bc-security')) . '" rel="noreferrer">' . esc_html__('Errors should never be printed', 'bc-security') . '</a>',
+            '<a href="' . esc_url(__('https://make.wordpress.org/core/handbook/testing/reporting-security-vulnerabilities/#why-are-there-path-disclosures-when-directly-loading-certain-files', 'bc-security')) . '" rel="noreferrer">' . esc_html__('display of errors can lead to path disclosures', 'bc-security') . '</a>'
         );
+    }
+
+
+    public function getName(): string
+    {
+        return __('Display of PHP errors is off', 'bc-security');
     }