+++ classes/BlueChip/Security/Modules/Checklist/Checks/NoAccessToPhpFilesInUploadsDirectory.php
@@ -8,16 +8,19 @@ use BlueChip\Security\Modules\Checklist;
 
 class NoAccessToPhpFilesInUploadsDirectory extends Checklist\BasicCheck
 {
-    public function __construct()
+    public function getDescription(): string
     {
-        parent::__construct(
-            __('No access to PHP files in uploads directory', 'bc-security'),
-            \sprintf(
-                /* translators: 1: link to gist with .htaccess configuration that disables access to PHP files */
-                esc_html__('Vulnerable plugins may allow upload of arbitrary files into uploads directory. %1$s within uploads directory may help prevent successful exploitation of such vulnerabilities.', 'bc-security'),
-                '<a href="https://gist.github.com/chesio/8f83224840eccc1e80a17fc29babadf2" rel="noreferrer">' . esc_html__('Disabling access to PHP files', 'bc-security') . '</a>'
-            )
+        return \sprintf(
+            /* translators: 1: link to gist with .htaccess configuration that disables access to PHP files */
+            esc_html__('Vulnerable plugins may allow upload of arbitrary files into uploads directory. %1$s within uploads directory may help prevent successful exploitation of such vulnerabilities.', 'bc-security'),
+            '<a href="https://gist.github.com/chesio/8f83224840eccc1e80a17fc29babadf2" rel="noreferrer">' . esc_html__('Disabling access to PHP files', 'bc-security') . '</a>'
         );
+    }
+
+
+    public function getName(): string
+    {
+        return __('No access to PHP files in uploads directory', 'bc-security');
     }