+++ classes/BlueChip/Security/Modules/Checklist/Checks/NoMd5HashedPasswords.php
@@ -15,16 +15,37 @@ class NoMd5HashedPasswords extends Checklist\BasicCheck
     private const WP_OLD_HASH_PREFIX = '$P$';
 
 
+    public function getDescription(): string
+    {
+        return \sprintf(
+            /* translators: 1: link to plugin with alternative implementation of password hashing scheme */
+            esc_html__('WordPress by default uses an MD5 based password hashing scheme that is too cheap and fast to generate cryptographically secure hashes. For modern PHP versions, there are %1$s available.', 'bc-security'),
+            '<a href="https://github.com/roots/wp-password-bcrypt" rel="noreferrer">' . esc_html__('more secure alternatives', 'bc-security') . '</a>'
+        );
+    }
+
+
+    public function getName(): string
+    {
+        return __('No default MD5 password hashes', 'bc-security');
+    }
+
+
     public function __construct(private wpdb $wpdb)
     {
-        parent::__construct(
-            __('No default MD5 password hashes', 'bc-security'),
-            \sprintf(
-                /* translators: 1: link to plugin with alternative implementation of password hashing scheme */
-                esc_html__('WordPress by default uses an MD5 based password hashing scheme that is too cheap and fast to generate cryptographically secure hashes. For modern PHP versions, there are %1$s available.', 'bc-security'),
-                '<a href="https://github.com/roots/wp-password-bcrypt" rel="noreferrer">' . esc_html__('more secure alternatives', 'bc-security') . '</a>'
-            )
-        );
+    }
+
+
+    /**
+     * Check does not make sense anymore starting with WordPress 6.8
+     *
+     * @link https://make.wordpress.org/core/2025/02/17/wordpress-6-8-will-use-bcrypt-for-password-hashing/
+     *
+     * @return bool
+     */
+    public function isMeaningful(): bool
+    {
+        return !is_wp_version_compatible('6.8');
     }