+++ classes/BlueChip/Security/Modules/Checklist/Checks/PluginsIntegrity.php
@@ -16,17 +16,20 @@ class PluginsIntegrity extends Checklist\AdvancedCheck
     protected const CRON_JOB_HOOK = Jobs::PLUGINS_INTEGRITY_CHECK;
 
 
-    public function __construct()
+    public function getDescription(): string
     {
-        parent::__construct(
-            __('Plugin files are untouched', 'bc-security'),
-            \sprintf(
-                /* translators: 1: link to Wikipedia article about md5sum, 2: link to Plugins Directory at WordPress.org */
-                esc_html__('By comparing %1$s of local plugin files with checksums provided by WordPress.org it is possible to determine whether any of plugin files have been modified or there are any unknown files in plugin directories. Note that this check works only with plugins installed from %2$s.', 'bc-security'),
-                '<a href="' . esc_url(__('https://en.wikipedia.org/wiki/Md5sum', 'bc-security')) . '" rel="noreferrer">' . esc_html__('MD5 checksums', 'bc-security') . '</a>',
-                '<a href="' . esc_url(__('https://wordpress.org/plugins/', 'bc-security')) . '" rel="noreferrer">' . esc_html__('Plugins Directory', 'bc-security') . '</a>'
-            )
+        return \sprintf(
+            /* translators: 1: link to Wikipedia article about md5sum, 2: link to Plugins Directory at WordPress.org */
+            esc_html__('By comparing %1$s of local plugin files with checksums provided by WordPress.org it is possible to determine whether any of plugin files have been modified or there are any unknown files in plugin directories. Note that this check works only with plugins installed from %2$s.', 'bc-security'),
+            '<a href="' . esc_url(__('https://en.wikipedia.org/wiki/Md5sum', 'bc-security')) . '" rel="noreferrer">' . esc_html__('MD5 checksums', 'bc-security') . '</a>',
+            '<a href="' . esc_url(__('https://wordpress.org/plugins/', 'bc-security')) . '" rel="noreferrer">' . esc_html__('Plugins Directory', 'bc-security') . '</a>'
         );
+    }
+
+
+    public function getName(): string
+    {
+        return __('Plugin files are untouched', 'bc-security');
     }