+++ classes/BlueChip/Security/Modules/Notifications/AdminPage.php
@@ -42,7 +42,6 @@ class AdminPage extends AbstractPage
                 __('You have set <code>BC_SECURITY_MUTE_NOTIFICATIONS</code> to true, therefore all notifications are muted.', 'bc-security'),
                 AdminNotices::INFO,
                 false, // ~ not dismissible
-                false // ~ do not escape HTML
             );
         }
     }