+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # BC Security Changelog
 
+## Version 0.26.0 (2025-04-17)
+
+### Added
+
+* Plugin has been tested with WordPress 6.8 [#168](https://github.com/chesio/bc-security/issues/168).
+
+### Changed
+
+* Separate plugin and theme update notification is sent for each available plugin and theme update [#167](https://github.com/chesio/bc-security/issues/167). Legacy behaviour can be enabled with following filters: `bc-security/filter:all-plugin-updates-in-one-notification` for plugins and `bc-security/filter:all-theme-updates-in-one-notification` for themes.
+* Translations loading has been improved in order to avoid translations being loaded too early [#169](https://github.com/chesio/bc-security/issues/169).
+
 ## Version 0.25.0 (2024-10-28)
 
 This release has been tested with PHP 8.4.